Overhauled the parser and generators to allow tags, improve type-safety. (#3)
This adds tags inside inline comment blocks to define how certain arguments
should be handled in the generated ctypes bindings. An obvious one is to
convert treat pointers as void* so we can just pass around addresses rather
than trying to coerce ourselves into the ctypes pointer type.

Another tag indicates that a pointer is derived from a NumPy array. In this
case, we perform (run-time) type checks for incoming arguments to avoid errors
in the function call. We also check that the array is contiguous (which is 
what is usually expected from C++), though this can be disabled if desired.